# server/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LifeSaver AI OS API",
    "version": "1.0.0",
    "description": "Comprehensive API Reference and Command Center for LifeSaver AI productivity platform. Fully synced with the Node/Express backend."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Primary API Gateway"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your secure JWT token. Authenticated requests require this header format: `Authorization: Bearer <token>`."
      }
    },
    "schemas": {
...
      "AIProcessRequest": {
        "type": "object",
        "properties": {
          "context": { "type": "string", "example": "I need to complete coding tasks, sleep 8 hours, and prepare for exams." },
          "tasks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "title": { "type": "string", "example": "DSA Revision" },
                "dueDate": { "type": "string", "example": "2026-06-30" },
                "urgency": { "type": "string", "example": "high" }
              }
            }
          },
          "mode": { "type": "string", "example": "general" }
        }
      },
      "AINegotiateRequest": {
        "type": "object",
        "required": ["deadlineTitle", "recipientName", "reason"],
        "properties": {
          "deadlineTitle": { "type": "string", "example": "React Architecture Document" },
          "recipientName": { "type": "string", "example": "Professor H. Miller" },
          "reason": { "type": "string", "example": "Suffered from an internet downtime outage for 48 hours" },
          "tone": { "type": "string", "enum": ["professional", "casual", "urgent"], "example": "professional" }
        }
      },
      "AITwinChatRequest": {
        "type": "object",
        "required": ["userMessage"],
        "properties": {
          "userMessage": { "type": "string", "example": "I feel unmotivated today, how do I get back on track?" },
          "context": { "type": "string", "example": "Wants to complete 2 DSA assignments before dark." },
          "chatHistory": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "role": { "type": "string", "example": "user" },
                "content": { "type": "string", "example": "Hello, when is my deadline?" }
              }
            }
          }
        }
      },
      "AIPrepareRequest": {
        "type": "object",
        "required": ["examOrInterviewTitle", "timeRemaining"],
        "properties": {
          "examOrInterviewTitle": { "type": "string", "example": "Fullstack Coding Interview" },
          "timeRemaining": { "type": "string", "example": "3 days" },
          "topicsToCover": {
            "type": "array",
            "items": { "type": "string", "example": "System Design, Sharding, Load Balancers" }
          }
        }
      }
    }
  },
  "paths": {
...
    "/ai/process": {
      "post": {
        "summary": "Run active Gemini-driven planning simulation engine",
        "tags": ["AI Core Operations"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AIProcessRequest" } }
          }
        },
        "responses": {
          "200": { "description": "Optimized schedule timelines, burnout risk forecasts, and multi-agent reports" }
        }
      }
    },
    "/ai/negotiate": {
      "post": {
        "summary": "Draft a persuasive, automated deadline extension request",
        "tags": ["AI Core Operations"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AINegotiateRequest" } }
          }
        },
        "responses": {
          "200": { "description": "Extension draft complete with custom tone" }
        }
      }
    },
    "/ai/twin-chat": {
      "post": {
        "summary": "Simulate dialogue with your productivity future-self twin",
        "tags": ["AI Core Operations"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AITwinChatRequest" } }
          }
        },
        "responses": {
          "200": { "description": "Dialog feedback and micro-actions returned" }
        }
      }
    },
    "/ai/prepare": {
      "post": {
        "summary": "Generate structural phase-study programs for upcoming exam/interview",
        "tags": ["AI Core Operations"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AIPrepareRequest" } }
          }
        },
        "responses": {
          "200": { "description": "Study timeline program" }
        }
      }
    },
...
  }
}